+++ app/info_tab.py
@@ -1,0 +1,5 @@
+from libraries import * 
+
+
+def info_tab():
+    st.write("# hello world")
