--- hello.py
@@ -1,7 +1,0 @@
-import streamlit as st
-import pandas as pd
-import numpy as np
-
-
-def hello_tab():
-    st.write("# hello world")
