+++ script.sh
@@ -1,0 +1,4 @@
+apt update
+apt upgrade -y
+cd /app
+pip install -r requirements.txt
